Split DecaExpr related functions to new file. Create util.py.
