+++ mira/sources/acsets/decapodes/__init__.py
@@ -1,1 +1,2 @@
 from .decapodes import *
+from .deca_expr import *
